refactor: implement feature-based MVVM folder structure
