+++ app/src/main/java/com/example/financial_broccoli/core/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.financial_broccoli.ui.theme
+package com.example.financial_broccoli.core.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
