+++ app/src/main/java/com/example/financial_broccoli/feature_expense/ExpenseViewModelFactory.kt
@@ -1,8 +1,8 @@
-package com.example.financial_broccoli.viewmodel
+package com.example.financial_broccoli.feature_expense
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.example.financial_broccoli.data.ExpenseRepository
+import com.example.financial_broccoli.feature_expense.data.ExpenseRepository
 
 class ExpenseViewModelFactory(private val repository: ExpenseRepository): ViewModelProvider.Factory {
