+++ app/src/main/java/com/example/financial_broccoli/feature_expense/data/ExpenseDatabase.kt
@@ -1,4 +1,4 @@
-package com.example.financial_broccoli.data
+package com.example.financial_broccoli.feature_expense.data
 
 import android.content.Context
 import androidx.room.Database
@@ -6,18 +6,18 @@ import androidx.room.Room
 import androidx.room.RoomDatabase
 
 @Database(entities = [Expense::class], version = 1)
-abstract class AppDatabase : RoomDatabase() {
+abstract class ExpenseDatabase : RoomDatabase() {
     abstract fun expenseDao(): ExpenseDao
 
     companion object {
         @Volatile
-        private var INSTANCE: AppDatabase? = null
+        private var INSTANCE: ExpenseDatabase? = null
 
-        fun getInstance(context: Context): AppDatabase {
+        fun getInstance(context: Context): ExpenseDatabase {
             return INSTANCE ?: synchronized(this) {
                 val instance = Room.databaseBuilder(
                     context.applicationContext,
-                    AppDatabase::class.java,
+                    ExpenseDatabase::class.java,
                     "database-name"
                 )
                     .fallbackToDestructiveMigration(true)
