+++ app/src/main/java/com/example/financial_broccoli/feature_expense/ui/ExpensesScreen.kt
@@ -1,4 +1,4 @@
-package com.example.financial_broccoli.ui
+package com.example.financial_broccoli.feature_expense.ui
 
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -17,7 +17,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.livedata.observeAsState
-import androidx.compose.runtime.mutableDoubleStateOf
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
@@ -26,7 +25,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
-import com.example.financial_broccoli.viewmodel.ExpenseViewModel
+import com.example.financial_broccoli.feature_expense.ExpenseViewModel
 
 @Composable
 fun ExpensesScreen(viewModel: ExpenseViewModel) {
